pyproject.toml
[project]
name = "datalake-serverless"
version = "0.1.0"
description = ""
authors = [
    {name = "VinicinCanedo",email = "vinicanedo@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "boto3 (>=1.39.3,<2.0.0)",
]

[tool.poetry]
packages = [{include = "datalake_serverless", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
